st-image-partitions.inc: Avoid noisy warning for empty partitions
Partitions like vendorfs can be in many cases empty due to not deploying
package targeting that specific partition - see
gcnano-userland-multi-binary-stm32mp as an example. Having a warning for
use-cases that are both common but also valid, would just result in
build noise. This change switches to build notes for flagging an empty
partition.

Signed-off-by: Andrei Gherzan <andrei.gherzan@huawei.com>